Refactor - Adicionando components para simplificar a lógica do arquivo principal
